add game modes and enemy layer

+++ Assets/Scripts/NPC/NpcStats.cs
@@ -9,7 +9,7 @@ public class NpcStats : MonoBehaviour
     public int power;
     public int laneIndex;
 
-    [SerializeField] private int currentHealth;
+    [SerializeField] private float currentHealth;
     [SerializeField] private LaneManager laneManager;
 
     protected virtual void Start()
@@ -20,8 +20,8 @@ public class NpcStats : MonoBehaviour
     public virtual void DoDamage(NpcStats _targetStats)
     {
 
-        int totalDamage = damage.GetValue();
-        _targetStats.TakeDamage(totalDamage);
+        float totalDamage = damage.GetValue();
+        _targetStats.TakeDamage((int)totalDamage);
     }
 
     public virtual void TakeDamage(int _damage)
